requirements.txt
# Core dependencies
PyYAML>=6.0.0

# Add other dependencies as needed
# python-telegram-bot>=20.0
# requests>=2.25.0
# aiohttp>=3.8.0